+++ src/main/java/ac/knu/likeknu/controller/dto/shuttlebus/ShuttleBusesArrivalTimeResponse.java
@@ -1,0 +1,4 @@
+package ac.knu.likeknu.controller.dto.shuttlebus;
+
+public class ShuttleBusesArrivalTimeResponse {
+}
